Segregated SideMenuController from global.dart
