+++ lib/easy_sidemenu.dart
@@ -1,6 +1,6 @@
 library easy_sidemenu;
 
-export 'package:easy_sidemenu/src/global/global.dart' show SideMenuController;
+export 'package:easy_sidemenu/src/side_menu_controller.dart';
 export 'package:easy_sidemenu/src/side_menu.dart';
 export 'package:easy_sidemenu/src/side_menu_display_mode.dart';
 export 'package:easy_sidemenu/src/side_menu_item.dart';
