+++ lib/src/side_menu_item.dart
@@ -1,7 +1,7 @@
 import 'package:badges/badges.dart' as bdg;
 import 'package:flutter/material.dart';
 import 'package:easy_sidemenu/src/side_menu_display_mode.dart';
-import 'global/global.dart';
+import 'package:easy_sidemenu/src/side_menu_controller.dart';
 
 class SideMenuItem {
   /// Title text
