+++ lib/src/side_menu_item_with_global.dart
@@ -2,6 +2,7 @@ import 'package:badges/badges.dart' as bdg;
 import 'package:flutter/material.dart';
 import 'package:easy_sidemenu/src/side_menu_display_mode.dart';
 import 'global/global.dart';
+import 'package:easy_sidemenu/src/side_menu_controller.dart';
 
 typedef SideMenuItemBuilder = Widget Function(
     BuildContext context, SideMenuDisplayMode displayMode);
