+++ test/easy_sidemenu_test.dart
@@ -1,5 +1,6 @@
 import 'package:easy_sidemenu/src/global/global.dart';
 import 'package:easy_sidemenu/src/side_menu.dart';
+import 'package:easy_sidemenu/src/side_menu_controller.dart';
 import 'package:easy_sidemenu/src/side_menu_display_mode.dart';
 import 'package:easy_sidemenu/src/side_menu_item.dart';
 import 'package:easy_sidemenu/src/side_menu_style.dart';
